to make separate folder structure for types,interface and scss
