+++ angular.json
@@ -112,7 +112,7 @@
             "styles": [
               "@angular/material/prebuilt-themes/azure-blue.css",
               "src/styles.scss",
-              "src/app/utils/_variable.scss"
+              "src/assets/scss/utils/_variable.scss"
             ]
           }
         },
