+++ src/app/shared/components/tag/tag.component.scss
@@ -1,106 +1,106 @@
-@use "../../../utils/_variable.scss";
+@use '../../../../assets/scss/utils/_variable.scss';
 
-@each $color-name in "black", "white", "green", "light-green", "brown", "light-brown", "purple", "light-purple", "yellow", "light-yellow" {
-    .bg-#{$color-name} {
-        --bg-color: var(--#{$color-name}-color);
-    }
+@each $color-name in 'black', 'white', 'green', 'light-green', 'brown', 'light-brown', 'purple',
+  'light-purple', 'yellow', 'light-yellow'
+{
+  .bg-#{$color-name} {
+    --bg-color: var(--#{$color-name}-color);
+  }
 
-    .text-#{$color-name} {
-        --text-color: var(--#{$color-name}-color);
-    }
+  .text-#{$color-name} {
+    --text-color: var(--#{$color-name}-color);
+  }
 }
 
 .tag-container {
-    display: inline-block;
+  display: inline-block;
 }
 
 .not-closable {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .bordered {
-    border: 1.5px solid var(--text-color) !important;
+  border: 1.5px solid var(--text-color) !important;
 }
 
 .tag {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  padding: 2px 8px;
+  border-radius: 12px;
+  font-weight: 500;
+  font-size: 11px;
+  line-height: 1.1;
+  position: relative;
+  overflow: hidden;
+  user-select: none;
+  min-height: 22px;
+  box-sizing: border-box;
+  background-color: var(--bg-color);
+  color: var(--text-color);
+  transition: all 0.15s ease-in-out;
+
+  &.static {
+    pointer-events: none;
+    .close-button {
+      pointer-events: none;
+      cursor: default;
+    }
+  }
+
+  .tag-close {
+    cursor: pointer;
+  }
+
+  .close-button {
     display: inline-flex;
     align-items: center;
-    gap: 4px;
-    padding: 2px 8px;
-    border-radius: 12px;
-    font-weight: 500;
-    font-size: 11px;
-    line-height: 1.1;
-    position: relative;
-    overflow: hidden;
-    user-select: none;
-    min-height: 22px;
-    box-sizing: border-box;
+    justify-content: center;
+    width: 14px;
+    height: 14px;
+    margin-left: 4px;
+    padding: 0;
+    border: none;
+    border-radius: 50%;
     background-color: var(--bg-color);
-    color: var(--text-color);
+    color: inherit;
     transition: all 0.15s ease-in-out;
 
-    &.static {
-        pointer-events: none;
-        .close-button {
-            pointer-events: none;
-            cursor: default;
-        }
+    svg {
+      width: 12px;
+      height: 12px;
+      stroke-width: 1.5;
     }
+  }
 
+  @media (max-width: 768px) {
+    .tag-chip {
+      padding: 1px 2px;
+      font-size: 8px;
 
-
-    .tag-close {
-        cursor: pointer;
-    }
-
-    .close-button {
-        display: inline-flex;
-        align-items: center;
-        justify-content: center;
-        width: 14px;
-        height: 14px;
-        margin-left: 4px;
-        padding: 0;
-        border: none;
-        border-radius: 50%;
-        background-color: var(--bg-color);
-        color: inherit;
-        transition: all 0.15s ease-in-out;
+      .close-button {
+        width: 10px;
+        height: 10px;
 
         svg {
-            width: 12px;
-            height: 12px;
-            stroke-width: 1.5;
+          width: 6px;
+          height: 6px;
         }
+      }
     }
-
-    @media (max-width: 768px) {
-        .tag-chip {
-            padding: 1px 2px;
-            font-size: 8px;
-
-            .close-button {
-                width: 10px;
-                height: 10px;
-
-                svg {
-                    width: 6px;
-                    height: 6px;
-                }
-            }
-        }
-    }
+  }
 }
 
 @keyframes fadeOut {
-    from {
-        opacity: 1;
-        transform: scale(1);
-    }
+  from {
+    opacity: 1;
+    transform: scale(1);
+  }
 
-    to {
-        opacity: 0.7;
-        transform: scale(1);
-    }
+  to {
+    opacity: 0.7;
+    transform: scale(1);
+  }
 }
