+++ src/app/shared/interfaces/tag.interface.ts
@@ -1,0 +1,12 @@
+import { TagColor } from '../../utils/types/tag-color.type';
+import { TagType } from '../components/tag/tag.component';
+
+export interface TagInputConfig {
+  id: string;
+  label: string;
+  type: TagType;
+  isSelected: boolean;
+  hasBorder: boolean;
+  backgroundColor: TagColor;
+  textColor: TagColor;
+}
