--- src/app/utils/_variable.scss
@@ -1,13 +1,0 @@
-:root {
-    --black-color: #0f172a;
-    --white-color: #f8fafc;
-    --green-color: #166534;
-    --brown-color: #c16a6a;
-    --purple-color: #6b21a8;
-    --yellow-color: #854d0e;
-
-    --light-green-color: #dcfce7;
-    --light-brown-color: #fee2e2;
-    --light-purple-color: #f3e8ff;
-    --light-yellow-color: #fef9c3;
-}
