+++ src/app/utils/types/tag-color.type.ts
@@ -1,0 +1,11 @@
+export type TagColor =
+  | 'black'
+  | 'white'
+  | 'green'
+  | 'brown'
+  | 'purple'
+  | 'yellow'
+  | 'lightGreen'
+  | 'lightBrown'
+  | 'lightPurple'
+  | 'lightYellow';
