+++ src/app/utils/types/tag.type.ts
@@ -1,0 +1,1 @@
+export type TagType = 'selectable' | 'static';
